move more operations in types

+++ kernel-lib/src/process/mod.rs
@@ -4,7 +4,7 @@ pub fn exit(code: i32) -> ! {
     unsafe {
         core::arch::asm! {
             "int 80",
-            in("eax") syscall::Request::TerminateCurrentTask,
+            in("eax") syscall::Request::TerminateCurrentTask as u32,
             in("edx") code,
             options(noreturn)
         }
